requirements.txt

clams-python==1.3.3
librosa>=0.10.1
soundfile>=0.12
ffmpeg-python>=0.2.0
torch>=2.1.0
speechbrain>=0.5.16
numpy
